feat(admin-ui): change uma resource detail page to modal
