+++ admin-ui/app/locales/pt/translation.json
@@ -78,7 +78,8 @@
     "remove": "Remover",
     "test": "Teste",
     "view": "Visualizar",
-    "yes": "Sim"
+    "yes": "Sim",
+    "close": "Fechar"
   },
   "fields": {
     "access_token_signing_alg": "Algoritmo de assinatura de token de acesso",
