+++ admin-ui/app/styles/custom/layout.scss
@@ -118,3 +118,8 @@
 .common-link {
   color: #2767b3
 }
+
+.top-5 {
+  position: relative;
+  top: 5px;
+}
