+++ admin-ui/plugins/auth-server/plugin-metadata.js
@@ -5,7 +5,6 @@ import ScopeEditPage from './components/Scopes/ScopeEditPage'
 import ClientListPage from './components/Clients/ClientListPage'
 import ClientAddPage from './components/Clients/ClientAddPage'
 import ClientEditPage from './components/Clients/ClientEditPage'
-import ClientUMADetailPage from './components/Clients/ClientUMADetailPage'
 
 import PropertiesPage from './components/Configuration/ConfigPage'
 import KeysPage from './components/Configuration/Keys/KeysPage'
@@ -91,11 +90,6 @@ const pluginMetadata = {
     {
       component: ClientEditPage,
       path: PLUGIN_BASE_APTH + '/client/edit:id',
-      permission: CLIENT_WRITE,
-    },
-    {
-      component: ClientUMADetailPage,
-      path: PLUGIN_BASE_APTH + '/client/uma/:id',
       permission: CLIENT_WRITE,
     },
     {
